requirements.txt
numpy==1.19.5
python-rtmidi==1.4.6
future==0.17.1
mido==1.2.9
PyAudio==0.2.11
matplotlib==3.3.3
librosa==0.8.0